Module bootstrap::builder

Re-exports

pub use crate::Compiler;

Structs

Builder

Cargo

RunConfig

Rustflags 🔒

Represents flag values in String form with whitespace delimiter to pass it to the compiler later.

ShouldRun

StepDescription 🔒

TaskPath

Enums

Alias

Kind

PathSet

Collection of paths used to match a task rule.

ReallyDefault 🔒

Traits

Step

Functions

crate_description

A description of the crates in this set, suitable for passing to builder.info.