Module bootstrap::compile

Implementation of compiling various phases of the compiler and standard library.

This module contains some of the real meat in the rustbuild build system which is where Cargo is used to compile the standard library, libtest, and the compiler. This module is also responsible for assembling the sysroot as it goes along from the output of the previous stage.

Structs

Assemble

CargoTarget

CodegenBackend

Rustc

RustcLink 🔒

StartupObjects

Std

StdLink 🔒

Sysroot

Enums

CargoMessage

Constants

CODEGEN_BACKEND_PREFIX 🔒

Functions

add_to_sysroot

Link some files into a rustc sysroot.

apple_darwin_sign_file 🔒

apple_darwin_update_library_name 🔒

codegen_backend_stamp 🔒

Cargo’s output path for librustc_codegen_llvm in a given stage, compiled by a particular compiler for the specified target and backend.

compiler_file

copy_and_stamp 🔒

copy_codegen_backends_to_sysroot 🔒

Creates the codegen-backends folder for a compiler that’s about to be assembled as a complete compiler.

copy_llvm_libunwind 🔒

copy_sanitizers 🔒

Copies sanitizer runtime libraries into target libdir.

copy_self_contained_objects 🔒

Copies third party objects needed by various targets for self-contained linkage.

copy_third_party_objects 🔒

Copies third party objects needed by various targets.

cp_rustc_component_to_ci_sysroot 🔒

is_codegen_cfg_needed 🔒

librustc_stamp

Cargo’s output path for librustc in a given stage, compiled by a particular compiler for the specified target.

libstd_stamp

Cargo’s output path for the standard library in a given stage, compiled by a particular compiler for the specified target.

needs_codegen_config 🔒

run_cargo

rustc_cargo

rustc_cargo_env

rustc_llvm_env 🔒

Pass down configuration from the LLVM build into the build of rustc_llvm and rustc_codegen_llvm.

std_cargo

Configure cargo to compile the standard library, adding appropriate env vars and such.

stream_cargo

strip_debug