Module bootstrap::config

Serialized configuration of a build.

This module implements parsing config.toml configuration files to tweak how the build runs.

Re-exports

pub use crate::flags::Subcommand;

Macros

check_ci_llvm 🔒

define_config 🔒

Structs

Build 🔒

TOML representation of various global build decisions.

CompilerMetadata

Config

Global configuration for the entire build and/or bootstrap.

Dist 🔒

Install 🔒

TOML representation of various global install decisions.

Llvm 🔒

TOML representation of how the LLVM build is configured.

OptimizeVisitor 🔒

Rust 🔒

TOML representation of how the Rust build is configured.

RustfmtMetadata

Stage0Config

Stage0Metadata

Target

Per-target configuration stored in the global configuration structure.

TargetSelection

TargetSelectionList

Newtype over Vec<TargetSelection> so we can implement custom parsing logic

TomlConfig 🔒

Structure of the config.toml file that configuration is read from.

TomlTarget 🔒

TOML representation of how each build target is configured.

Enums

DebuginfoLevel

DryRun

LlvmLibunwind

ReplaceOpt 🔒

Describes how to handle conflicts in merging two TomlConfig

RustOptimize

RustcLto

LTO mode used for compiling rustc itself.

RustfmtState

SplitDebuginfo

StringOrBool

StringOrInt 🔒

Traits

Merge 🔒

Functions

format_optimize_error_msg 🔒

set 🔒

target_selection_list

threads_from_config 🔒