Module bootstrap::core::build_steps::clean

./x.py clean

Responsible for cleaning out a build directory of all old and stale artifacts to prepare for a fresh build. Currently doesn’t remove the build/cache directory (download cache) or the build/$target/llvm directory unless the --all flag is present.

Macros

clean_crate_tree 🔒

Structs

CleanAll

Rustc

Std

Functions

clean 🔒

clean_default 🔒

clean_specific_stage 🔒

do_op 🔒

rm_rf 🔒