bootstrap::core::build_steps

Module doc

Documentation generation for bootstrap.

This module implements generation for all bits and pieces of documentation for the Rust project. This notably includes suites like the rust book, the nomicon, rust by example, standalone documentation, etc.

Everything here is basically just a shim around calling either rustbook or rustdoc.

Macros

book 🔒

tool_doc 🔒

Structs

Bootstrap

BuildHelper

Cargo

CargoBook

Clippy

ClippyBook

Compiletest

EditionGuide

EmbeddedBook

ErrorIndex

Miri

Nomicon

Reference

Documents the reference. It has to always be done using a stage 1+ compiler, because it references in-tree compiler/stdlib concepts.

Releases

RunMakeSupport

RustByExample

RustbookSrc 🔒

Rustc

Document the compiler for the given target using rustdoc from build_compiler.

RustcBook

Builds the Rust compiler book.

Rustdoc

RustdocBook

Rustfmt

SharedAssets

SharedAssetsPaths

Standalone

Std

Document the standard library using build_compiler.

StyleGuide

TheBook

Tidy

UnstableBook

UnstableBookGen

Enums

DocumentationFormat

Constants

STD_PUBLIC_CRATES 🔒

Name of the crates that are visible to consumers of the standard library. Documentation for internal crates is handled by the rustc step, so internal crates will show up there.

Functions

doc_std 🔒

Build the documentation for public standard library crates.

invoke_rustdoc 🔒

prepare_doc_compiler 🔒

Prepare a compiler that will be able to document something for target at stage.

symlink_dir_force 🔒