bootstrap::core::build_steps

Module run

Build-and-run steps for in-repo tools

A bit of a hodge-podge as e.g. if a tool’s a test fixture it should be in build_steps::test. If it can be reached from ./x.py run it can go here.

Structs

BuildManifest

BumpStage0

CollectLicenseMetadata

CoverageDump

Step to manually run the coverage-dump tool (./x run coverage-dump).

CyclicStep

Dummy step that can be used to deliberately trigger bootstrap’s step cycle detector, for automated and manual testing.

FeaturesStatusDump

GenerateCompletions

GenerateCopyright

GenerateWindowsSys

Miri

ReplaceVersionPlaceholder

Rustfmt

UnicodeTableGenerator

Functions

get_completion_paths

Return tuples of (shell, file containing completions).