Module bootstrap::core::build_steps::setup

First time setup of a dev environment

These are build-and-run steps for ./x.py setup, which allows quickly setting up the directory for modifying, building, and running the compiler and library. Running arbitrary configuration allows setting up things that cannot be simply captured inside the config.toml, in addition to leading people away from manually editing most of the config.toml values.

Structs

Hook

Installs src/etc/pre-push.sh as a Git hook

Link

Creates a toolchain link for stage1 using rustup

Vscode

Sets up or displays src/etc/rust_analyzer_settings.json

Enums

Profile

PromptResult 🔒

Statics

PROFILE_DIR 🔒

RUST_ANALYZER_SETTINGS 🔒

SETTINGS_HASHES 🔒

A list of historical hashes of src/etc/rust_analyzer_settings.json. New entries should be appended whenever this is updated so we can detect outdated vs. user-modified settings files.

Functions

attempt_toolchain_link 🔒

create_vscode_settings_maybe 🔒

Create a .vscode/settings.json file for rustc development, or just print it If this method should be re-called, it returns false.

ensure_stage1_toolchain_placeholder_exists 🔒

install_git_hook_maybe 🔒

interactive_path

prompt_user 🔒

Prompt a user for a answer, looping until they enter an accepted input or nothing

rustup_installed 🔒

setup

setup_config_toml 🔒

stage_dir_exists 🔒

toolchain_is_linked 🔒

try_link_toolchain 🔒