bootstrap::core::build_steps

Module vendor

Handles the vendoring process for the bootstrap system.

This module ensures that all required Cargo dependencies are gathered and stored in the <src>/<VENDOR_DIR> directory.

Structs

Vendor 🔒

Defines the vendoring step in the bootstrap process.

VendorOutput 🔒

Stores the result of the vendoring step.

Constants

VENDOR_DIR

The name of the directory where vendored dependencies are stored.

Functions

default_paths_to_vendor

Returns the cargo workspaces to vendor for x vendor and dist tarballs.