bootstrap::core

Module builder

Re-exports

pub use self::cargo::Cargo;

pub use self::cargo::cargo_profile_var;

pub use crate::Compiler;

Modules

cargo 🔒

Structs

Builder

Builds and performs different Self::kinds of stuff and actions, taking into account build configuration from e.g. bootstrap.toml.

CLIStepPath 🔒

Libdir 🔒

RunConfig

ShouldRun

StepDescription 🔒

StepMetadata

Metadata that describes an executed step, mostly for testing and tracing.

TaskPath

Enums

Alias

Kind

PathSet

Collection of paths used to match a task rule.

ReallyDefault 🔒

Constants

PATH_REMAP 🔒

Traits

AnyDebug

This trait is similar to Any, except that it also exposes the underlying type’s Debug implementation.

Step

Functions

crate_description

A description of the crates in this set, suitable for passing to builder.info.

remap_paths 🔒