bootstrap::core::builder
Struct Builder

pub struct Builder<'a> {
    pub build: &'a Build,
    pub top_stage: u32,
    pub kind: Kind,
    cache: Cache,
    stack: RefCell<Vec<Box<dyn Any>>>,
    time_spent_on_dependencies: Cell<Duration>,
    pub paths: Vec<PathBuf>,
}

Builds and performs different Self::kinds of stuff and actions, taking into account build configuration from e.g. config.toml.

Fields
build: &'a Build

Build configuration from e.g. config.toml.

top_stage: u32

The stage to use. Either implicitly determined based on subcommand, or explicitly specified with --stage N. Normally this is the stage we use, but sometimes we want to run steps with a lower stage than this.

kind: Kind

What to build or what action to perform.

cache: Cache

A cache of outputs of Steps so we can avoid running steps we already ran.

stack: RefCell<Vec<Box<dyn Any>>>

A stack of Steps to run before we can run this builder. The output of steps is cached in Self::cache.

time_spent_on_dependencies: Cell<Duration>

The total amount of time we spent running Steps in Self::stack.

paths: Vec<PathBuf>

The paths passed on the command line. Used by steps to figure out what to do. For example: with ./x check foo bar we get paths=["foo", "bar"].

Implementations

impl Builder<'_>

pub(crate) fn msg_tool( &self, kind: Kind, mode: Mode, tool: &str, build_stage: u32, host: &TargetSelection, target: &TargetSelection, ) -> Option<Group>

impl<'a> Builder<'a>

pub fn tool_exe(&self, tool: Tool) -> PathBuf

impl Builder<'_>

pub fn tool_cmd(&self, tool: Tool) -> BootstrapCommand

Gets a BootstrapCommand which is ready to run tool in stage built for host.

impl Builder<'_>

fn toolstates(&self) -> HashMap<Box<str>, ToolState>

pub fn save_toolstate(&self, tool: &str, state: ToolState)

Updates the actual toolstate of a tool.

The toolstates are saved to the file specified by the key rust.save-toolstates in config.toml. If unspecified, nothing will be done. The file is updated immediately after this function completes.

impl Builder<'_>

pub fn bare_cargo( &self, compiler: Compiler, mode: Mode, target: TargetSelection, cmd_kind: Kind, ) -> BootstrapCommand

Like Builder::cargo, but only passes flags that are valid for all commands.

fn cargo( &self, compiler: Compiler, mode: Mode, source_type: SourceType, target: TargetSelection, cmd_kind: Kind, ) -> Cargo

This will create a BootstrapCommand that represents a pending execution of cargo. This cargo will be configured to use compiler as the actual rustc compiler, its output will be scoped by mode’s output directory, it will pass the --target flag for the specified target, and will be executing the Cargo command cmd. cmd can be miri-cmd for commands to be run with Miri.

impl<'a> Builder<'a>

fn get_step_descriptions(kind: Kind) -> Vec<StepDescription>

pub fn get_help(build: &Build, kind: Kind) -> Option<String>

fn new_internal(build: &Build, kind: Kind, paths: Vec<PathBuf>) -> Builder<'_>

pub fn new(build: &Build) -> Builder<'_>

pub fn execute_cli(&self)

pub fn default_doc(&self, paths: &[PathBuf])

pub fn doc_rust_lang_org_channel(&self) -> String

fn run_step_descriptions(&self, v: &[StepDescription], paths: &[PathBuf])

pub fn link_std_into_rustc_driver(&self, target: TargetSelection) -> bool

Returns if std should be statically linked into rustc_driver. It’s currently not done on windows-gnu due to linker bugs.

pub fn compiler(&self, stage: u32, host: TargetSelection) -> Compiler

Obtain a compiler at a given stage and for a given host (i.e., this is the target that the compiler will run on, not the target it will build code for). Explicitly does not take Compiler since all Compiler instances are meant to be obtained through this function, since it ensures that they are valid (i.e., built and assembled).

pub fn compiler_for( &self, stage: u32, host: TargetSelection, target: TargetSelection, ) -> Compiler

Similar to compiler, except handles the full-bootstrap option to silently use the stage1 compiler instead of a stage2 compiler if one is requested.

Note that this does not have the side effect of creating compiler(stage, host), unlike compiler above which does have such a side effect. The returned compiler here can only be used to compile new artifacts, it can’t be used to rely on the presence of a particular sysroot.

See force_use_stage1 and force_use_stage2 for documentation on what each argument is.

pub fn sysroot(&self, compiler: Compiler) -> PathBuf

pub fn sysroot_target_bindir( &self, compiler: Compiler, target: TargetSelection, ) -> PathBuf

Returns the bindir for a compiler’s sysroot.

pub fn sysroot_target_libdir( &self, compiler: Compiler, target: TargetSelection, ) -> PathBuf

Returns the libdir where the standard library and other artifacts are found for a compiler’s sysroot.

pub fn sysroot_codegen_backends(&self, compiler: Compiler) -> PathBuf

pub fn rustc_libdir(&self, compiler: Compiler) -> PathBuf

Returns the compiler’s libdir where it stores the dynamic libraries that it itself links against.

For example this returns <sysroot>/lib on Unix and <sysroot>/bin on Windows.

pub fn libdir_relative(&self, compiler: Compiler) -> &Path

Returns the compiler’s relative libdir where it stores the dynamic libraries that it itself links against.

For example this returns lib on Unix and bin on Windows.

pub fn sysroot_libdir_relative(&self, compiler: Compiler) -> &Path

Returns the compiler’s relative libdir where the standard library and other artifacts are found for a compiler’s sysroot.

For example this returns lib on Unix and Windows.

pub fn rustc_lib_paths(&self, compiler: Compiler) -> Vec<PathBuf>

pub fn add_rustc_lib_path(&self, compiler: Compiler, cmd: &mut BootstrapCommand)

Adds the compiler’s directory of dynamic libraries to cmd’s dynamic library lookup path.

pub fn rustc(&self, compiler: Compiler) -> PathBuf

Gets a path to the compiler specified.

fn codegen_backends(&self, compiler: Compiler) -> impl Iterator<Item = PathBuf>

Gets the paths to all of the compiler’s codegen backends.

pub fn rustdoc(&self, compiler: Compiler) -> PathBuf

pub fn cargo_clippy_cmd(&self, run_compiler: Compiler) -> BootstrapCommand

pub fn cargo_miri_cmd(&self, run_compiler: Compiler) -> BootstrapCommand

pub fn rustdoc_cmd(&self, compiler: Compiler) -> BootstrapCommand

fn llvm_config(&self, target: TargetSelection) -> Option<PathBuf>

Return the path to llvm-config for the target, if it exists.

Note that this returns None if LLVM is disabled, or if we’re in a check build or dry-run, where there’s no need to build all of LLVM.

pub fn ensure<S: Step>(&'a self, step: S) -> S::Output

Ensure that a given step is built, returning its output. This will cache the step, so it is safe (and good!) to call this as often as needed to ensure that all dependencies are built.

pub(crate) fn ensure_if_default<T, S: Step<Output = Option<T>>>( &'a self, step: S, kind: Kind, ) -> S::Output

Ensure that a given step is built only if it’s supposed to be built by default, returning its output. This will cache the step, so it’s safe (and good!) to call this as often as needed to ensure that all dependencies are build.

pub(crate) fn was_invoked_explicitly<S: Step>(&'a self, kind: Kind) -> bool

Checks if any of the “should_run” paths is in the Builder paths.

pub(crate) fn maybe_open_in_browser<S: Step>(&self, path: impl AsRef<Path>)

pub(crate) fn open_in_browser(&self, path: impl AsRef<Path>)