bootstrap::core::config::config

Function is_system_llvm

pub fn is_system_llvm<'a>(
    dwn_ctx: impl AsRef<DownloadContext<'a>>,
    target: TargetSelection,
) -> bool

Returns true if this is an external version of LLVM not managed by bootstrap. In particular, we expect llvm sources to be available when this is false.

NOTE: this is not the same as !is_rust_llvm when llvm_has_patches is set.