bootstrap::core::config::toml

Module llvm

This module defines the Llvm struct, which represents the [llvm] table in the bootstrap.toml configuration file.

Structs

Llvm

TOML representation of how the LLVM build is configured.

Functions

check_incompatible_options_for_ci_llvm

Compares the current Llvm options against those in the CI LLVM builder and detects any incompatible options. It does this by destructuring the Llvm instance to make sure every Llvm field is covered and not missing.