bootstrap::core::config::toml

Module rust

This module defines the Rust struct, which represents the [rust] table in the bootstrap.toml configuration file.

Structs

OptimizeVisitor 🔒

Rust

TOML representation of how the Rust build is configured.

Enums

LldMode

LLD in bootstrap works like this:

RustOptimize

Constants

BUILTIN_CODEGEN_BACKENDS 🔒

Functions

check_incompatible_options_for_ci_rustc

Compares the current Rust options against those in the CI rustc builder and detects any incompatible options. It does this by destructuring the Rust instance to make sure every Rust field is covered and not missing.

default_lld_opt_in_targets 🔒

format_optimize_error_msg 🔒

parse_codegen_backends 🔒