bootstrap::core
Module download

Enums

DownloadSource 🔒

Statics

SHOULD_FIX_BINS_AND_DYLIBS 🔒

Functions

curl_version 🔒

extract_curl_version 🔒

is_download_ci_available 🔒

Checks whether the CI rustc is available for the given target triple.

path_is_dylib 🔒

recorded_entries 🔒

try_run 🔒

Config::try_run wrapper for this module to avoid warnings on try_run, since we don’t have access to a builder yet.