Module bootstrap::core::metadata

Structs

Dependency 🔒

For more information, see the output of https://doc.rust-lang.org/nightly/cargo/commands/cargo-metadata.html

Output 🔒

For more information, see the output of https://doc.rust-lang.org/nightly/cargo/commands/cargo-metadata.html

Package 🔒

For more information, see the output of https://doc.rust-lang.org/nightly/cargo/commands/cargo-metadata.html

Target 🔒

Functions

build

Collects and stores package metadata of each workspace members into build, by executing cargo metadata commands.

workspace_members 🔒

Invokes cargo metadata to get package metadata of each workspace member.