Enum bootstrap::PathSet

pub enum PathSet {
    Set(BTreeSet<TaskPath>),
    Suite(TaskPath),
}

Collection of paths used to match a task rule.

Variants

Set(BTreeSet<TaskPath>)

A collection of individual paths or aliases.

These are generally matched as a path suffix. For example, a command-line value of std will match if library/std is in the set.

NOTE: the paths within a set should always be aliases of one another. For example, src/librustdoc and src/tools/rustdoc should be in the same set, but library/core and library/std generally should not, unless there’s no way (for that Step) to build them separately.

Suite(TaskPath)

A “suite” of paths.

These can match as a path suffix (like Set), or as a prefix. For example, a command-line value of tests/ui/abi/variadic-ffi.rs will match tests/ui. A command-line value of ui would also match tests/ui.

Implementations

impl PathSet

fn empty() -> PathSet

fn one<P: Into<PathBuf>>(path: P, kind: Kind) -> PathSet

fn has(&self, needle: &Path, module: Kind) -> bool

fn check(p: &TaskPath, needle: &Path, module: Kind) -> bool

fn intersection_removing_matches( &self, needles: &mut Vec<&Path>, module: Kind ) -> PathSet

Return all TaskPaths in Self that contain any of the needles, removing the matched needles.

This is used for StepDescription::krate, which passes all matching crates at once to Step::make_run, rather than calling it many times with a single crate. See tests.rs for examples.

pub fn assert_single_path(&self) -> &TaskPath

A convenience wrapper for Steps which know they have no aliases and all their sets contain only a single path.

This can be used with ShouldRun::crate_or_deps, ShouldRun::path, or ShouldRun::alias.