Enum bootstrap::Subcommand

pub enum Subcommand {

    Build,
    Check {
        all_targets: bool,
    },
    Clippy {
        fix: bool,
        allow: Vec<String>,
        deny: Vec<String>,
        warn: Vec<String>,
        forbid: Vec<String>,
    },
    Fix,
    Format {
        check: bool,
    },
    Doc {
        open: bool,
        json: bool,
    },
    Test {

        no_fail_fast: bool,
        skip: Vec<PathBuf>,
        test_args: Vec<String>,
        rustc_args: Vec<String>,
        no_doc: bool,
        doc: bool,
        bless: bool,
        extra_checks: Option<String>,
        force_rerun: bool,
        only_modified: bool,
        compare_mode: Option<String>,
        pass: Option<String>,
        run: Option<String>,
        rustfix_coverage: bool,
    
},
    Bench {
        test_args: Vec<String>,
    },
    Clean {
        all: bool,
        stage: Option<u32>,
    },
    Dist,
    Install,
    Run {
        args: Vec<String>,
    },
    Setup {
        profile: Option<PathBuf>,
    },
    Suggest {
        run: bool,
    },

}

Variants

Build

Compile either the compiler or libraries

Check

Fields

all_targets: bool

Check all targets

Compile either the compiler or libraries, using cargo check

Clippy

Fields

fix: bool

allow: Vec<String>

clippy lints to allow

deny: Vec<String>

clippy lints to deny

warn: Vec<String>

clippy lints to warn on

forbid: Vec<String>

clippy lints to forbid

Run Clippy (uses rustup/cargo-installed clippy binary)

Fix

Run cargo fix

Format

Fields

check: bool

check formatting instead of applying

Run rustfmt

Doc

Fields

open: bool

open the docs in a browser

json: bool

render the documentation in JSON format in addition to the usual HTML format

Build documentation

Test

Fields

no_fail_fast: bool

run all tests regardless of failure

skip: Vec<PathBuf>

skips tests matching SUBSTRING, if supported by test tool. May be passed multiple times

test_args: Vec<String>

extra arguments to be passed for the test tool being used (e.g. libtest, compiletest or rustdoc)

rustc_args: Vec<String>

extra options to pass the compiler when running tests

no_doc: bool

do not run doc tests

doc: bool

only run doc tests

bless: bool

whether to automatically update stderr/stdout files

extra_checks: Option<String>

comma-separated list of other files types to check (accepts py, py:lint, py:fmt, shell)

force_rerun: bool

rerun tests even if the inputs are unchanged

only_modified: bool

only run tests that result has been changed

compare_mode: Option<String>

mode describing what file the actual ui output will be compared to

pass: Option<String>

force {check,build,run}-pass tests to this mode.

run: Option<String>

whether to execute run-* tests

rustfix_coverage: bool

enable this to generate a Rustfix coverage file, which is saved in /<build_base>/rustfix_missing_coverage.txt

Build and run some test suites

Bench

Fields

test_args: Vec<String>

Build and run some benchmarks

Clean

Fields

all: bool

Clean the entire build directory (not used by default)

stage: Option<u32>

Clean a specific stage without touching other artifacts. By default, every stage is cleaned if this option is not used.

Clean out build directories

Dist

Build distribution artifacts

Install

Install distribution artifacts

Run

Fields

args: Vec<String>

arguments for the tool

Run tools contained in this repository

Setup

Fields

profile: Option<PathBuf>

Either the profile for config.toml or another setup action. May be omitted to set up interactively

Set up the environment for development

Suggest

Fields

run: bool

run suggested tests

Suggest a subset of tests to run, based on modified files

Implementations

impl Subcommand

pub fn kind(&self) -> Kind

pub fn rustc_args(&self) -> Vec<&str>

pub fn fail_fast(&self) -> bool

pub fn doc_tests(&self) -> DocTests

pub fn bless(&self) -> bool

pub fn extra_checks(&self) -> Option<&str>

pub fn only_modified(&self) -> bool

pub fn force_rerun(&self) -> bool

pub fn rustfix_coverage(&self) -> bool

pub fn compare_mode(&self) -> Option<&str>

pub fn pass(&self) -> Option<&str>

pub fn run(&self) -> Option<&str>

pub fn open(&self) -> bool

pub fn json(&self) -> bool