bootstrap
Struct Flags

pub struct Flags {

    pub cmd: Subcommand,
    pub verbose: u8,
    pub incremental: bool,
    pub config: Option<PathBuf>,
    pub build_dir: Option<PathBuf>,
    pub build: Option<String>,
    pub host: Option<TargetSelectionList>,
    pub target: Option<TargetSelectionList>,
    pub exclude: Vec<PathBuf>,
    pub skip: Vec<PathBuf>,
    pub include_default_paths: bool,
    pub rustc_error_format: Option<String>,
    pub on_fail: Option<String>,
    pub dry_run: bool,
    pub dump_bootstrap_shims: bool,
    pub stage: Option<u32>,
    pub keep_stage: Vec<u32>,
    pub keep_stage_std: Vec<u32>,
    pub src: Option<PathBuf>,
    pub jobs: Option<u32>,
    pub warnings: Warnings,
    pub error_format: Option<String>,
    pub json_output: bool,
    pub color: Color,
    pub bypass_bootstrap_lock: bool,
    pub rust_profile_generate: Option<String>,
    pub rust_profile_use: Option<String>,
    pub llvm_profile_use: Option<String>,
    pub llvm_profile_generate: bool,
    pub enable_bolt_settings: bool,
    pub skip_stage0_validation: bool,
    pub reproducible_artifact: Vec<String>,
    pub paths: Vec<PathBuf>,
    pub set: Vec<String>,
    pub free_args: Vec<String>,

}

Deserialized version of all flags for this compile.

Fields
cmd: Subcommandverbose: u8

use verbose output (-vv for very verbose)

incremental: bool

use incremental compilation

config: Option<PathBuf>

TOML configuration file for build

build_dir: Option<PathBuf>

Build directory, overrides build.build-dir in config.toml

build: Option<String>

build target of the stage0 compiler

host: Option<TargetSelectionList>

host targets to build

target: Option<TargetSelectionList>

target targets to build

exclude: Vec<PathBuf>

build paths to exclude

skip: Vec<PathBuf>

build paths to skip

include_default_paths: bool

include default paths in addition to the provided ones

rustc_error_format: Option<String>on_fail: Option<String>

command to run on failure

dry_run: bool

dry run; don’t build anything

dump_bootstrap_shims: bool

Indicates whether to dump the work done from bootstrap shims

stage: Option<u32>

stage to build (indicates compiler to use/test, e.g., stage 0 uses the bootstrap compiler, stage 1 the stage 0 rustc artifacts, etc.)

keep_stage: Vec<u32>

stage(s) to keep without recompiling (pass multiple times to keep e.g., both stages 0 and 1)

keep_stage_std: Vec<u32>

stage(s) of the standard library to keep without recompiling (pass multiple times to keep e.g., both stages 0 and 1)

src: Option<PathBuf>

path to the root of the rust checkout

jobs: Option<u32>

number of jobs to run in parallel

warnings: Warnings

if value is deny, will deny warnings if value is warn, will emit warnings otherwise, use the default configured behaviour

error_format: Option<String>

rustc error format

json_output: bool

use message-format=json

color: Color

whether to use color in cargo and rustc output

bypass_bootstrap_lock: bool

Bootstrap uses this value to decide whether it should bypass locking the build process. This is rarely needed (e.g., compiling the std library for different targets in parallel).

Unless you know exactly what you are doing, you probably don’t need this.

rust_profile_generate: Option<String>

generate PGO profile with rustc build

rust_profile_use: Option<String>

use PGO profile for rustc build

llvm_profile_use: Option<String>

use PGO profile for LLVM build

llvm_profile_generate: bool

generate PGO profile with llvm built for rustc

enable_bolt_settings: bool

Enable BOLT link flags

skip_stage0_validation: bool

Skip stage0 compiler validation

reproducible_artifact: Vec<String>

Additional reproducible artifacts that should be added to the reproducible artifacts archive.

paths: Vec<PathBuf>

paths for the subcommand

set: Vec<String>

override options in config.toml

free_args: Vec<String>

arguments passed to subcommands

Implementations

impl Flags

pub fn try_parse_verbose_help(args: &[String]) -> bool

Check if <cmd> -h -v was passed. If yes, print the available paths and return true.

pub fn parse(args: &[String]) -> Self