Module bootstrap::test

Implementation of the test-related targets of the build system.

This file implements the various regression test suites that we execute on our CI.

Macros

default_test 🔒

default_test_with_compare_mode 🔒

host_test 🔒

test 🔒

test_book 🔒

test_definitions 🔒

test_with_compare_mode 🔒

Structs

Assembly

BookTest 🔒

Bootstrap

Cargo

Cargotest

Clippy

Codegen

CodegenUnits

Compiletest 🔒

CompiletestTest

Crate

CrateBootstrap

CrateLibrustc

CrateRustdoc

Rustdoc is special in various ways, which is why this step is different from Crate.

CrateRustdocJsonTypes

Debuginfo

Distcheck

EditionGuide

EmbeddedBook

ErrorIndex

ExpandYamlAnchors

HtmlCheck

Incremental

Linkcheck

LintDocs

MirOpt

Miri

Nomicon

Pretty

Reference

RemoteCopyLibs

Some test suites are run inside emulators or on remote devices, and most of our test binaries are linked dynamically which means we need to ship the standard library and such to the emulator ahead of time. This step represents this and is a dependency of all test suites.

RunCoverage

RunCoverageRustdoc

RunMake

RunMakeFullDeps

RunPassValgrind

RustAnalyzer

RustByExample

RustDemangler

RustInstaller

RustcBook

RustcGuide

Rustdoc

RustdocBook

RustdocGUI

RustdocJSNotStd

RustdocJSStd

RustdocJson

RustdocTheme

RustdocUi

Rustfmt

TestHelpers

TheBook

Tidy

TierCheck

Ui

UiFullDeps

UnstableBook

Constants

ADB_TEST_DIR 🔒

MIR_OPT_BLESS_TARGET_MAPPING 🔒

Functions

check_if_tidy_is_installed 🔒

get_browser_ui_test_version 🔒

get_browser_ui_test_version_inner 🔒

markdown_test 🔒

path_for_cargo 🔒

prepare_cargo_test 🔒

Given a cargo test subcommand, pass it the appropriate test flags given a builder.

run_cargo_test 🔒

Given a cargo test subcommand, add the appropriate flags and run it.

testdir 🔒