Module bootstrap::util

Various utility functions used throughout rustbuild.

Simple things like testing the various filesystem operations here and there, not a lot of interesting happenings here unfortunately.

Macros

t

A helper macro to unwrap a result except also print out details like:

Structs

TimeIt

Functions

absolute 🔒

Copied from std::path::absolute until it stabilizes.

absolute_unix 🔒

Make a POSIX path absolute without changing its semantics.

add_dylib_path

Adds a list of lookup paths to cmd’s dynamic library lookup path. If the dylib_path_var is already set for this cmd, the old value will be overwritten!

add_link_lib_path

Adds a list of lookup paths to cmd’s link library lookup path.

check_run

dir_is_empty

dir_up_to_date 🔒

dylib_path

Parses the dylib_path_var() environment variable, returning a list of paths that are members of this lookup path.

dylib_path_var

Returns the environment variable which the dynamic library lookup path resides in for this platform.

exe

Given an executable called name, return the filename for the executable for a particular target.

forcing_clang_based_tests

get_clang_cl_resource_dir

Adapted from https://github.com/llvm/llvm-project/blob/782e91224601e461c019e0a4573bbccc6094fbcd/llvm/cmake/modules/HandleLLVMOptions.cmake#L1058-L1079

is_debug_info

Returns true if the file name given looks like a debug info file

is_dylib

Returns true if the file name given looks like a dynamic library.

is_valid_test_suite_arg

libdir

Returns the corresponding relative library directory that the compiler’s dylibs will be found in.

link_lib_path 🔒

Parses the link_lib_path_var() environment variable, returning a list of paths that are members of this lookup path.

link_lib_path_var 🔒

Returns the environment variable which the link library lookup path resides in for this platform.

lld_flag_no_threads

make

mtime

Returns the last-modified time for path, or zero if it doesn’t exist.

output

output_result

program_out_of_date 🔒

Used for download caching

run

run_suppressed

symlink_dir

Symlinks two directories, using junctions on Windows and normal symlinks on Unix.

timeit

Returns an RAII structure that prints out how long it took to drop.

try_run_suppressed

up_to_date

Returns true if dst is up to date given that the file or files in src are used to generate it.

use_host_linker