bootstrap

Module utils

This module contains integral components of the build and configuration process, providing support for a wide range of tasks and operations such as caching, tarballs, release channels, job management, etc.

Modules

build_stamp 🔒

Module for managing build stamp files.

cache 🔒

This module helps you efficiently store and retrieve values using interning.

cc_detect 🔒

C-compiler probing and detection.

change_tracker 🔒

This module facilitates the tracking system for major changes made to the bootstrap, with the goal of keeping developers synchronized with important modifications in the bootstrap.

channel 🔒

Build configuration for Rust’s release channels.

exec 🔒

Command Execution Module

helpers 🔒

Various utility functions used throughout bootstrap.

job 🔒

render_tests 🔒

This module renders the JSON output of libtest into a human-readable form, trying to be as similar to libtest’s native output as possible.

shared_helpers 🔒

This module serves two purposes:

tarball 🔒

Facilitates the management and generation of tarballs.

tracing 🔒

Wrapper macros for tracing macros to avoid having to write cfg(feature = "tracing")-gated debug!/trace! everytime, e.g.