bootstrap::utils

Module build_stamp

Module for managing build stamp files.

Contains the core implementation of how bootstrap utilizes stamp files on build processes.

Structs

BuildStamp

Manages a stamp file to track build state. The file is created in the given directory and can have custom content and name.

Functions

clear_if_dirty

Clear out dir if input is newer.

codegen_backend_stamp

Cargo’s output path for librustc_codegen_llvm in a given stage, compiled by a particular compiler for the specified target and backend.

generate_smart_stamp_hash

Computes a hash representing the state of a repository/submodule and additional input.

librustc_stamp

Cargo’s output path for librustc in a given stage, compiled by a particular compiler for the specified target.

libstd_stamp

Cargo’s output path for the standard library in a given stage, compiled by a particular compiler for the specified target.