bootstrap::utils::cache

Trait Internable

trait Internable:
    Clone
    + Eq
    + Hash
    + 'static {
    // Required method
    fn intern_cache() -> &'static Mutex<TyIntern<Self>>;

    // Provided method
    fn intern(self) -> Interned<Self> { ... }
}

Defines the behavior required for a type to be internable.

Types implementing this trait must provide access to a static cache and define an intern method that ensures values are stored uniquely.

Required Methods

fn intern_cache() -> &'static Mutex<TyIntern<Self>>

Provided Methods

fn intern(self) -> Interned<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.