bootstrap::utils::channel
Enum GitInfo

pub enum GitInfo {
    Absent,
    Present(Option<Info>),
    RecordedForTarball(Info),
}

Variants

Absent

This is not a git repository.

Present(Option<Info>)

This is a git repository. If the info should be used (omit_git_hash is false), this will be Some, otherwise it will be None.

RecordedForTarball(Info)

This is not a git repository, but the info can be fetched from the git-commit-info file.

Implementations

impl GitInfo

pub fn new(omit_git_hash: bool, dir: &Path) -> GitInfo

pub fn info(&self) -> Option<&Info>

pub fn sha(&self) -> Option<&str>

pub fn sha_short(&self) -> Option<&str>

pub fn commit_date(&self) -> Option<&str>

pub fn version(&self, build: &Build, num: &str) -> String

pub fn is_managed_git_subrepository(&self) -> bool

Returns whether this directory has a .git directory which should be managed by bootstrap.

pub fn is_from_tarball(&self) -> bool

Returns whether this is being built from a tarball.