bootstrap::utils::exec

Struct BootstrapCommand

pub struct BootstrapCommand {
    command: Command,
    pub failure_behavior: BehaviorOnFailure,
    pub run_in_dry_run: bool,
    drop_bomb: DropBomb,
    should_cache: bool,
}

Wrapper around std::process::Command.

By default, the command will exit bootstrap if it fails. If you want to allow failures, use allow_failure. If you want to delay failures until the end of bootstrap, use delay_failure.

By default, the command will print its stdout/stderr to stdout/stderr of bootstrap (OutputMode::Print). If you want to handle the output programmatically, use BootstrapCommand::run_capture.

Bootstrap will print a debug log to stdout if the command fails and failure is not allowed.

By default, command executions are cached based on their workdir, program, arguments, and environment variables. This avoids re-running identical commands unnecessarily, unless caching is explicitly disabled.

Fields
command: Commandfailure_behavior: BehaviorOnFailurerun_in_dry_run: booldrop_bomb: DropBombshould_cache: bool
Implementations

impl<'a> BootstrapCommand

pub fn new<S: AsRef<OsStr>>(program: S) -> Self

pub fn arg<S: AsRef<OsStr>>(&mut self, arg: S) -> &mut Self

pub fn cached(&mut self) -> &mut Self

Cache the command. If it will be executed multiple times with the exact same arguments and environment variables in the same bootstrap invocation, the previous result will be loaded from memory.

pub fn args<I, S>(&mut self, args: I) -> &mut Self
where I: IntoIterator<Item = S>, S: AsRef<OsStr>,

pub fn env<K, V>(&mut self, key: K, val: V) -> &mut Self
where K: AsRef<OsStr>, V: AsRef<OsStr>,

pub fn get_envs(&self) -> CommandEnvs<'_>

pub fn get_args(&self) -> CommandArgs<'_>

pub fn env_remove<K: AsRef<OsStr>>(&mut self, key: K) -> &mut Self

pub fn current_dir<P: AsRef<Path>>(&mut self, dir: P) -> &mut Self

pub fn stdin(&mut self, stdin: Stdio) -> &mut Self

pub fn delay_failure(self) -> Self

pub fn fail_fast(self) -> Self

pub fn allow_failure(self) -> Self

pub fn run_in_dry_run(&mut self) -> &mut Self

pub fn run(&mut self, exec_ctx: impl AsRef<ExecutionContext>) -> bool

Run the command, while printing stdout and stderr. Returns true if the command has succeeded.

pub fn run_capture( &mut self, exec_ctx: impl AsRef<ExecutionContext>, ) -> CommandOutput

Run the command, while capturing and returning all its output.

pub fn run_capture_stdout( &mut self, exec_ctx: impl AsRef<ExecutionContext>, ) -> CommandOutput

Run the command, while capturing and returning stdout, and printing stderr.

pub fn start_capture( &'a mut self, exec_ctx: impl AsRef<ExecutionContext>, ) -> DeferredCommand<'a>

Spawn the command in background, while capturing and returning all its output.

pub fn start_capture_stdout( &'a mut self, exec_ctx: impl AsRef<ExecutionContext>, ) -> DeferredCommand<'a>

Spawn the command in background, while capturing and returning stdout, and printing stderr.

pub fn stream_capture_stdout( &'a mut self, exec_ctx: impl AsRef<ExecutionContext>, ) -> Option<StreamingCommand>

Spawn the command in background, while capturing and returning stdout, and printing stderr. Returns None in dry-mode

pub fn mark_as_executed(&mut self)

Mark the command as being executed, disarming the drop bomb. If this method is not called before the command is dropped, its drop will panic.

pub fn get_created_location(&self) -> Location<'static>

Returns the source code location where this command was created.

pub fn force_coloring_in_ci(&mut self)

If in a CI environment, forces the command to run with colors.

pub fn fingerprint(&self) -> CommandFingerprint