bootstrap::utils::exec

Struct ExecutionContext

pub struct ExecutionContext {
    dry_run: DryRun,
    pub verbosity: u8,
    pub fail_fast: bool,
    delayed_failures: Arc<Mutex<Vec<String>>>,
    command_cache: Arc<CommandCache>,
    profiler: Arc<CommandProfiler>,
}

Fields
dry_run: DryRunverbosity: u8fail_fast: booldelayed_failures: Arc<Mutex<Vec<String>>>command_cache: Arc<CommandCache>profiler: Arc<CommandProfiler>
Implementations

impl ExecutionContext

pub fn new(verbosity: u8, fail_fast: bool) -> Self

pub fn dry_run(&self) -> bool

pub fn profiler(&self) -> &CommandProfiler

pub fn get_dry_run(&self) -> &DryRun

pub fn verbose(&self, f: impl Fn())

pub fn is_verbose(&self) -> bool

pub fn fail_fast(&self) -> bool

pub fn set_dry_run(&mut self, value: DryRun)

pub fn set_verbosity(&mut self, value: u8)

pub fn set_fail_fast(&mut self, value: bool)

pub fn add_to_delay_failure(&self, message: String)

pub fn report_failures_and_exit(&self)

pub fn start<'a>( &self, command: &'a mut BootstrapCommand, stdout: OutputMode, stderr: OutputMode, ) -> DeferredCommand<'a>

Execute a command and return its output. Note: Ideally, you should use one of the BootstrapCommand::run* functions to execute commands. They internally call this method.

pub fn run( &self, command: &mut BootstrapCommand, stdout: OutputMode, stderr: OutputMode, ) -> CommandOutput

Execute a command and return its output. Note: Ideally, you should use one of the BootstrapCommand::run* functions to execute commands. They internally call this method.

fn fail(&self, message: &str) -> !

pub fn stream( &self, command: &mut BootstrapCommand, stdout: OutputMode, stderr: OutputMode, ) -> Option<StreamingCommand>

Spawns the command with configured stdout and stderr handling.

Returns None if in dry-run mode or Panics if the command fails to spawn.