bootstrap::core::build_steps
Module clippy

Implementation of running clippy on the compiler, standard library and various tools.

Macros

lint_any 🔒

Structs

Bootstrap

BuildHelper

BuildManifest

CI

CargoMiri

Clippy

CollectLicenseMetadata

Compiletest

CoverageDump

Jsondocck

Jsondoclint

LintConfig

LintDocs

LlvmBitcodeLinker

Miri

MiroptTestTools

OptDist

RemoteTestClient

RemoteTestServer

Rls

RustAnalyzer

RustInstaller

Rustc

Rustdoc

Rustfmt

Std

TestFloatParse

Tidy

Constants

IGNORED_RULES_FOR_STD_AND_RUSTC 🔒

Disable the most spammy clippy lints

Functions

get_clippy_rules_in_order

We need to keep the order of the given clippy lint rules before passing them. Since clap doesn’t offer any useful interface for this purpose out of the box, we have to handle it manually.

lint_args 🔒