bootstrap::core::build_steps
Module doc

Documentation generation for bootstrap.

This module implements generation for all bits and pieces of documentation for the Rust project. This notably includes suites like the rust book, the nomicon, rust by example, standalone documentation, etc.

Everything here is basically just a shim around calling either rustbook or rustdoc.

Macros

book 🔒

submodule_helper 🔒

tool_doc 🔒

Structs

Bootstrap

BuildHelper

Cargo

CargoBook

Clippy

ClippyBook

Compiletest

EditionGuide

EmbeddedBook

ErrorIndex

Miri

Nomicon

Reference

Releases

RunMakeSupport

RustByExample

RustbookSrc 🔒

Rustc

RustcBook

Rustdoc

RustdocBook

Rustfmt

SharedAssets

SharedAssetsPaths

Standalone

Std

StyleGuide

TheBook

Tidy

UnstableBook

UnstableBookGen

Enums

DocumentationFormat

Constants

STD_PUBLIC_CRATES 🔒

Name of the crates that are visible to consumers of the standard library. Documentation for internal crates is handled by the rustc step, so internal crates will show up there.

Functions

doc_std 🔒

Build the documentation for public standard library crates.

invoke_rustdoc 🔒

symlink_dir_force 🔒