bootstrap::core::build_steps
Module install

Implementation of the install aspects of the compiler.

This module is responsible for installing the standard library, compiler, and documentation.

Macros

install 🔒

Structs

Cargo

Clippy

Docs

LlvmBitcodeLinker

LlvmTools

Miri

RustAnalyzer

Rustc

RustcCodegenCranelift

Rustfmt

Src

Std

Constants

SHELL 🔒

Functions

default_path 🔒

install_sh 🔒

is_dir_writable_for_user 🔒

prepare_dir 🔒

sanitize_sh 🔒

We have to run a few shell scripts, which choke quite a bit on both \ characters and on C:\ paths, so normalize both of them away.