bootstrap::core::build_steps
Module setup

First time setup of a dev environment

These are build-and-run steps for ./x.py setup, which allows quickly setting up the directory for modifying, building, and running the compiler and library. Running arbitrary configuration allows setting up things that cannot be simply captured inside the config.toml, in addition to leading people away from manually editing most of the config.toml values.

Structs

Editor

Sets up or displays the LSP config for one of the supported editors

Hook

Installs src/etc/pre-push.sh as a Git hook

Link

Creates a toolchain link for stage1 using rustup

Enums

EditorKind 🔒

Handles editor-specific setup differences

Profile

PromptResult 🔒

Statics

PROFILE_DIR 🔒

Functions

attempt_toolchain_link 🔒

create_editor_settings_maybe 🔒

Create the recommended editor LSP config file for rustc development, or just print it If this method should be re-called, it returns false.

ensure_stage1_toolchain_placeholder_exists 🔒

install_git_hook_maybe 🔒

interactive_path

prompt_user 🔒

Prompt a user for a answer, looping until they enter an accepted input or nothing

rustup_installed 🔒

setup

setup_config_toml 🔒

stage_dir_exists 🔒

toolchain_is_linked 🔒

try_link_toolchain 🔒