bootstrap::core::build_steps
Module test

Build-and-run steps for ./x.py test test fixtures

./x.py test (aka Kind::Test) is currently allowed to reach build steps in other modules. However, this contains ~all test parts we expect people to be able to build and run locally.

Macros

coverage_test_alias 🔒

Declares an alias for running the Coverage tests in only one mode. Adapted from test_definitions.

default_test 🔒

default_test_with_compare_mode 🔒

host_test 🔒

test 🔒

test_book 🔒

test_definitions 🔒

test_with_compare_mode 🔒

Structs

Assembly

BookTest 🔒

Bootstrap

Cargo

CargoMiri

Cargotest

Clippy

Codegen

CodegenCranelift

CodegenGCC

CodegenUnits

CollectLicenseMetadata

Compiletest 🔒

CompiletestTest

Coverage

Coverage tests are a bit more complicated than other test suites, because we want to run the same set of test files in multiple different modes, in a way that’s convenient and flexible when invoked manually.

CoverageMap

CoverageRun

CoverageRunRustdoc

Crashes

Crate

CrateBootstrap

CrateBuildHelper

CrateLibrustc

CrateRunMakeSupport

CrateRustdoc

Rustdoc is special in various ways, which is why this step is different from Crate.

CrateRustdocJsonTypes

Debuginfo

Distcheck

EditionGuide

EmbeddedBook

ErrorIndex

HtmlCheck

Incremental

Linkcheck

LintDocs

MirOpt

Miri

Nomicon

Pretty

Reference

RemoteCopyLibs

Some test suites are run inside emulators or on remote devices, and most of our test binaries are linked dynamically which means we need to ship the standard library and such to the emulator ahead of time. This step represents this and is a dependency of all test suites.

RunMake

Special-handling is needed for run-make, so don’t use default_test for defining RunMake tests.

RunMakeSupport

RustAnalyzer

RustByExample

RustInstaller

RustcBook

RustcGuide

Rustdoc

RustdocBook

RustdocGUI

RustdocJSNotStd

RustdocJSStd

RustdocJson

RustdocTheme

RustdocUi

Rustfmt

TestFloatParse

TestHelpers

TheBook

Tidy

TierCheck

Ui

UiFullDeps

UnstableBook

Constants

ADB_TEST_DIR 🔒

Functions

check_if_tidy_is_installed 🔒

get_browser_ui_test_version 🔒

get_browser_ui_test_version_inner 🔒

markdown_test 🔒

path_for_cargo 🔒

prepare_cargo_test 🔒

Given a cargo test subcommand, pass it the appropriate test flags given a builder.

run_cargo_test 🔒

Given a cargo test subcommand, add the appropriate flags and run it.

testdir 🔒