bootstrap::core::build_steps
Module tool

Macros

bootstrap_tool 🔒

tool_extended 🔒

Structs

BuildManifest

BumpStage0

Cargo

CargoClippy

CargoMiri

CargoTest

Cargofmt

Clippy

CollectLicenseMetadata

Compiletest

CoverageDump

ErrorIndex

GenerateCopyright

GenerateWindowsSys

HtmlChecker

JsonDocCk

JsonDocLint

LibcxxVersionTool

Linkchecker

LintDocs

LldWrapper

LlvmBitcodeLinker

Miri

OptimizedDist

RemoteTestClient

RemoteTestServer

ReplaceVersionPlaceholder

Rls

RustAnalyzer

RustAnalyzerProcMacroSrv

RustInstaller

Rustbook

RustcPerf

The rustc-perf benchmark suite, which is added as a submodule at src/tools/rustc-perf.

RustcPerfWrapper

Rustdoc

RustdocGUITest

RustdocTheme

Rustfmt

SuggestTests

TestFloatParse

Tidy

ToolBuild 🔒

UnicodeTableGenerator

UnstableBookGen

WasmComponentLd

Enums

LibcxxVersion

SourceType

Tool

Statics

SUBMODULES_FOR_RUSTBOOK

These are the submodules that are required for rustbook to work due to depending on mdbook plugins.

Functions

copy_link_tool_bin 🔒

Links a built tool binary with the given name from the build directory to the tools directory.

prepare_tool_cargo