bootstrap::core
Module builder

Re-exports

pub use self::cargo::Cargo;

pub use crate::Compiler;

Modules

cargo 🔒

Structs

Builder

Builds and performs different Self::kinds of stuff and actions, taking into account build configuration from e.g. config.toml.

Libdir 🔒

RunConfig

ShouldRun

StepDescription 🔒

TaskPath

Enums

Alias

Kind

PathSet

Collection of paths used to match a task rule.

ReallyDefault 🔒

Constants

PATH_REMAP 🔒

Traits

Step

Functions

crate_description

A description of the crates in this set, suitable for passing to builder.info.

remap_paths 🔒