bootstrap::core::config

Module flags

Command-line interface of the bootstrap build system.

This module implements the command-line parsing of the build system which has various flags to configure how it’s run.

Structs

Flags

Deserialized version of all flags for this compile.

Enums

Color

Subcommand

Warnings

Whether to deny warnings, emit them as warnings, or use the default behavior

Functions

get_completion

Returns the shell completion for a given shell, if the result differs from the current content of path. If path does not exist, always returns Some.

normalize_args 🔒