bootstrap::core

Module metadata

This module interacts with Cargo metadata to collect and store information about the packages in the Rust workspace.

It runs cargo metadata to gather details about each package, including its name, source, dependencies, targets, and available features. The collected metadata is then used to update the Build structure, ensuring proper dependency resolution and compilation flow.

Structs

Dependency 🔒

For more information, see the output of https://doc.rust-lang.org/nightly/cargo/commands/cargo-metadata.html

Output 🔒

For more information, see the output of https://doc.rust-lang.org/nightly/cargo/commands/cargo-metadata.html

Package 🔒

For more information, see the output of https://doc.rust-lang.org/nightly/cargo/commands/cargo-metadata.html

Functions

build

Collects and stores package metadata of each workspace members into build, by executing cargo metadata commands.

workspace_members 🔒

Invokes cargo metadata to get package metadata of each workspace member.