bootstrap
Module utils

This module contains integral components of the build and configuration process, providing support for a wide range of tasks and operations such as caching, tarballs, release channels, job management, etc.

Modules

cache 🔒

cc_detect 🔒

C-compiler probing and detection.

change_tracker 🔒

This module facilitates the tracking system for major changes made to the bootstrap, with the goal of keeping developers synchronized with important modifications in the bootstrap.

channel 🔒

Build configuration for Rust’s release channels.

exec 🔒

helpers 🔒

Various utility functions used throughout bootstrap.

job 🔒

render_tests 🔒

This module renders the JSON output of libtest into a human-readable form, trying to be as similar to libtest’s native output as possible.

shared_helpers 🔒

This module serves two purposes: 1. It is part of the utils module and used in other parts of bootstrap. 2. It is embedded inside bootstrap shims to avoid a dependency on the bootstrap library. Therefore, this module should never use any other bootstrap module. This reduces binary size and improves compilation time by minimizing linking time.

tarball 🔒

Facilitates the management and generation of tarballs.