bootstrap::utils
Module cache

Structs

Cache

This is essentially a HashMap which allows storing any type in its input and any type in its output. It is a write-once cache; values are never evicted, which means that references to the value can safely be returned from the get() method.

Interned

Interner

TyIntern 🔒

Statics

INTERNER

Traits

Internable 🔒